REEB-53
git-svn-id: http://svn.uhuila.cn/reeb/trunk@183 96626eaa-6fa6-4472-8f5b-47c901fe3adc

+++ module/order/app/models/order/OrderItems.java
@@ -7,6 +7,7 @@ import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.ManyToOne;
 import javax.persistence.Table;
+import java.math.BigDecimal;
 import java.util.Date;
 import java.util.List;
 
@@ -20,10 +21,10 @@ public class OrderItems extends Model {
     Goods goods;
 
     @Column(name="originalPrice")
-    Float originalPrice;
+    BigDecimal originalPrice;
 
     @Column(name="salePrice")
-    Float salePrice;
+    BigDecimal salePrice;
 
     @Column(name="goods_name")
     String goodsName;
